# project/tests/goldens/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Computational Metaphysics Engine",
    "description": "Modular 10-Domain Metaphysical Calculation, AI Debate & Multi-Agent Engine",
    "version": "1.0.0.34833bc"
  },
  "paths": {
...
    "/api/v1/calendar/month": {
      "get": {
        "tags": [
          "Astrological Calendar & Date Selection"
        ],
        "summary": "Get Monthly Calendar",
        "description": "Retrieve full monthly astrological calendar metadata with 12 duty officers and suitability.",
        "operationId": "get_monthly_calendar_api_v1_calendar_month_get",
        "parameters": [
          {
            "name": "year",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "description": "Calendar Year",
              "default": 2026,
              "title": "Year"
            },
            "description": "Calendar Year"
          },
          {
            "name": "month",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "description": "Calendar Month (1-12)",
              "default": 8,
              "title": "Month"
            },
            "description": "Calendar Month (1-12)"
          },
          {
            "name": "user_day_master",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "User Day Master",
              "title": "User Day Master"
            },
            "description": "User Day Master"
          },
          {
            "name": "user_zodiac",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "User Zodiac Branch",
              "title": "User Zodiac"
            },
            "description": "User Zodiac Branch"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Get Monthly Calendar Api V1 Calendar Month Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/calendar/query-dates": {
      "post": {
        "tags": [
          "Astrological Calendar & Date Selection"
        ],
        "summary": "Query Auspicious Dates",
        "description": "Find and rank auspicious dates for specific activity in upcoming period.",
        "operationId": "query_auspicious_dates_api_v1_calendar_query_dates_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DateQueryRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "additionalProperties": true,
                    "type": "object"
                  },
                  "type": "array",
                  "title": "Response Query Auspicious Dates Api V1 Calendar Query Dates Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/telegram/webhook": {
      "post": {
        "tags": [
          "Telegram Controller"
        ],
        "summary": "Telegram Webhook",
        "description": "Handles incoming Telegram updates & admin commands (Decision 2).",
        "operationId": "telegram_webhook_api_v1_telegram_webhook_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": true,
                "type": "object",
                "title": "Payload"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "DateQueryRequest": {
        "properties": {
          "intent": {
            "type": "string",
            "title": "Intent",
            "description": "business_opening | marriage_ceremony | home_moving | contract_signing | travel_journey | wealth_investment",
            "default": "business_opening"
          },
          "start_date": {
            "type": "string",
            "title": "Start Date",
            "description": "YYYY-MM-DD format"
          },
          "days_ahead": {
            "type": "integer",
            "title": "Days Ahead",
            "description": "Number of days to search ahead",
            "default": 30
          },
          "user_day_master": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "User Day Master",
            "description": "User's Day Master Stem (e.g. 甲, 乙)"
          },
          "user_zodiac": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "User Zodiac",
            "description": "User's Birth Year Branch (e.g. 子, 丑)"
          }
        },
        "type": "object",
        "required": [
          "start_date"
        ],
        "title": "DateQueryRequest"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}